# apps/cockpit/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckResultOut": {
        "properties": {
          "actual_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Value"
          },
          "duration_ms": {
            "default": 0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "expect": {
            "default": "",
            "title": "Expect",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "passed": {
            "default": false,
            "title": "Passed",
            "type": "boolean"
          },
          "severity": {
            "default": "fail",
            "title": "Severity",
            "type": "string"
          },
          "sql": {
            "default": "",
            "title": "Sql",
            "type": "string"
          },
          "state": {
            "default": "executed",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CheckResultOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObjectDetailOut": {
        "properties": {
          "check_count": {
            "default": 0,
            "title": "Check Count",
            "type": "integer"
          },
          "contract": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract"
          },
          "contract_status": {
            "default": "",
            "title": "Contract Status",
            "type": "string"
          },
          "cov_flag": {
            "default": "out_of_scope",
            "title": "Cov Flag",
            "type": "string"
          },
          "family": {
            "default": "quality",
            "title": "Family",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run"
          },
          "last_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run Id"
          },
          "latest_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunSummaryOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "layer": {
            "default": "",
            "title": "Layer",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owned_by": {
            "default": "platform",
            "title": "Owned By",
            "type": "string"
          },
          "run_history": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RunListItem"
            },
            "title": "Run History",
            "type": "array"
          },
          "schema_name": {
            "default": "",
            "title": "Schema Name",
            "type": "string"
          },
          "space": {
            "default": "",
            "title": "Space",
            "type": "string"
          },
          "status": {
            "default": "unknown",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "ObjectDetailOut",
        "type": "object"
      },
      "ObjectOut": {
        "properties": {
          "check_count": {
            "default": 0,
            "title": "Check Count",
            "type": "integer"
          },
          "contract_status": {
            "default": "",
            "title": "Contract Status",
            "type": "string"
          },
          "cov_flag": {
            "default": "out_of_scope",
            "title": "Cov Flag",
            "type": "string"
          },
          "family": {
            "default": "quality",
            "title": "Family",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run"
          },
          "last_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run Id"
          },
          "layer": {
            "default": "",
            "title": "Layer",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owned_by": {
            "default": "platform",
            "title": "Owned By",
            "type": "string"
          },
          "schema_name": {
            "default": "",
            "title": "Schema Name",
            "type": "string"
          },
          "space": {
            "default": "",
            "title": "Space",
            "type": "string"
          },
          "status": {
            "default": "unknown",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "ObjectOut",
        "type": "object"
      },
...
      "RunListItem": {
        "properties": {
          "dataset": {
            "title": "Dataset",
            "type": "string"
          },
          "failed": {
            "default": 0,
            "title": "Failed",
            "type": "integer"
          },
          "finished_at": {
            "default": "",
            "title": "Finished At",
            "type": "string"
          },
          "overall_status": {
            "default": "pass",
            "title": "Overall Status",
            "type": "string"
          },
          "passed": {
            "default": 0,
            "title": "Passed",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "run_state": {
            "default": "finished",
            "title": "Run State",
            "type": "string"
          },
          "started_at": {
            "default": "",
            "title": "Started At",
            "type": "string"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          },
          "triggered_by": {
            "default": "manual",
            "title": "Triggered By",
            "type": "string"
          },
          "warnings": {
            "default": 0,
            "title": "Warnings",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "dataset"
        ],
        "title": "RunListItem",
        "type": "object"
      },
      "RunSummaryOut": {
        "properties": {
          "actor": {
            "default": "",
            "title": "Actor",
            "type": "string"
          },
          "contract_hash": {
            "default": "",
            "title": "Contract Hash",
            "type": "string"
          },
          "contract_version": {
            "default": "",
            "title": "Contract Version",
            "type": "string"
          },
          "dataset": {
            "title": "Dataset",
            "type": "string"
          },
          "failed": {
            "default": 0,
            "title": "Failed",
            "type": "integer"
          },
          "finished_at": {
            "default": "",
            "title": "Finished At",
            "type": "string"
          },
          "overall_status": {
            "default": "pass",
            "title": "Overall Status",
            "type": "string"
          },
          "passed": {
            "default": 0,
            "title": "Passed",
            "type": "integer"
          },
          "results": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CheckResultOut"
            },
            "title": "Results",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "run_state": {
            "default": "finished",
            "title": "Run State",
            "type": "string"
          },
          "schema_name": {
            "default": "",
            "title": "Schema Name",
            "type": "string"
          },
          "started_at": {
            "default": "",
            "title": "Started At",
            "type": "string"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          },
          "triggered_by": {
            "default": "manual",
            "title": "Triggered By",
            "type": "string"
          },
          "warnings": {
            "default": 0,
            "title": "Warnings",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "dataset"
        ],
        "title": "RunSummaryOut",
        "type": "object"
      },
      "RunTriggerIn": {
        "properties": {
          "environment": {
            "default": "",
            "title": "Environment",
            "type": "string"
          },
          "execution_mode": {
            "default": "auto",
            "title": "Execution Mode",
            "type": "string"
          }
        },
        "title": "RunTriggerIn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "DQ & Observability Cockpit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/objects": {
      "get": {
        "operationId": "list_objects_api_objects_get",
        "parameters": [
          {
            "in": "query",
            "name": "space",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Space"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ObjectOut"
                  },
                  "title": "Response List Objects Api Objects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Objects",
        "tags": [
          "objects"
        ]
      }
    },
    "/api/objects/{object_id}": {
      "get": {
        "operationId": "get_object_api_objects__object_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Object",
        "tags": [
          "objects"
        ]
      }
    },
    "/api/objects/{object_id}/checks/{check_name}/history": {
      "get": {
        "description": "WS1-2: actual_value-Zeitreihe je Check \u2014 Sparkline- und Miner-Quelle.",
        "operationId": "get_check_history_api_objects__object_id__checks__check_name__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "check_name",
            "required": true,
            "schema": {
              "title": "Check Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Check History",
        "tags": [
          "objects"
        ]
      }
    },
    "/api/objects/{object_id}/run": {
      "post": {
        "description": "Trigger a DQ run for an object. Returns run_id immediately (202). [ENGINE-FROZEN]\n\n[AUTHZ] Runs l\u00f6sen Last auf HANA aus \u2014 viewer darf nicht triggern.\n[SCHEMA-MAP] '{schema}' aus kompilierten Checks wird HIER gebunden (G2):\naus dem Environment, sonst aus dem Inventar-Schema des Objekts.",
        "operationId": "trigger_run_api_objects__object_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunTriggerIn",
                "default": {
                  "environment": "",
                  "execution_mode": "auto"
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Run",
        "tags": [
          "objects"
        ]
      }
    },
    "/api/objects/{object_id}/runs": {
      "get": {
        "operationId": "get_object_runs_api_objects__object_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunListItem"
                  },
                  "title": "Response Get Object Runs Api Objects  Object Id  Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Object Runs",
        "tags": [
          "objects"
        ]
      }
    },
...
  }
}